+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Email templates for Likes and Reposts
 * Improve Interactions moderation
 * Compatibility with Akismet
+* Comment type mapping for `Like` and `Announce`
 
 ### Fixed